requirements.txt
pytest>=7.0.0
pyyaml>=6.0
wave>=0.0.2
click>=8.0.0
rich>=10.0.0
